api, app: 멤버 삭제 시 desk로 저장되도록 수정

+++ api/core/socketio/delete-location.js
@@ -8,10 +8,22 @@ const subscribeDeleteLocationEvent = (io, socket) => {
     //   return socket.emit(SOCKET_EVENTS.DELETE_LOCATION, { status: 401, message: "로그인이 필요합니다." });
     // }
 
-    const { roomId } = props;
+    const { data, roomId, location } = props;
+    if (data.type === "MEMBER") {
+      await partService.updateEmptyDeskMember({
+        roomId,
+        locationX: location.x,
+        locationY: location.y,
+      });
 
-    await partService.deleteLocation(props);
-    io.to(`room_${roomId}`).emit(SOCKET_EVENTS.DELETE_LOCATION, props);
+      props.data.type = "DESK";
+
+      io.to(`room_${roomId}`).emit(SOCKET_EVENTS.APPEND_LOCATION, props);
+    } else {
+      await partService.deleteLocation(props);
+
+      io.to(`room_${roomId}`).emit(SOCKET_EVENTS.DELETE_LOCATION, props);
+    }
   });
 };
